add address store

+++ src/ntp/ntp.c
@@ -23,7 +23,7 @@ int ntp_set_time_common(
         int timeout,
         int try_times) {
     wdt_feed();
-    time_t _build_t = (int)build_time();
+    time_t _build_t = build_time();
     int i=0;
     do {
         while( ntp_set_time(server, port, (uint32_t)timeout) != NTP_OK ) {
@@ -32,7 +32,7 @@ int ntp_set_time_common(
             if ( try_times >= 0 && i++ >= try_times ) return -1;
         }
 
-        DBG(" time diff %d %d ", (int)time(NULL), _build_t);
+        DBG(" time diff %d %d ", (int)time(NULL), (int)_build_t);
     } while(time(NULL) <= _build_t - 86400);
     return 0;
 }
